Finish code for data gathering

+++ Motor_Test/arduinoCode/DAQ_test/test.cpp
@@ -1,6 +1,6 @@
 #include <cxxopts.hpp>
 
-unsigned long long index = 0;
+double index = 0;
 
 int main() {
